parent BatchHandler
